api/requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
redis
ollama
pandas
pyodbc
cx-Oracle
python-jose[cryptography]
passlib[bcrypt]
python-ldap
python-multipart
PyPDF2
python-docx
openpyxl
langchain
sentence-transformers